improved support for larger samples
By using shutil.copyfileobj() instead of a plain .write(.read())
sequence we save the Cuckoo Agent from having to load the target file
completely in-memory. This greatly helps Cuckoo with larger samples as
VMs may not have been configured with the amount of memory required for
bigger samples (which could run up to, e.g., a few hundred MBs).